# web/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "High-Performance Dynamic API",
    "version": "1.0.0",
    "description": "L.S.D (Large Seach of Data) API with real-time search, CDC sync, cursor pagination, and ClickHouse/PostgreSQL support",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/tables/{table}/schema": {
      "get": {
        "tags": [
          "Table Discovery"
        ],
        "summary": "Get Table Schema",
        "description": "Get detailed column definitions for a specific table. Use to generate TypeScript types, validate inputs, or build dynamic forms.",
        "operationId": "getTableSchema",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Table name",
            "example": "b2b"
          }
        ],
        "responses": {
          "200": {
            "description": "Table schema retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/tables/{table}/records": {
      "get": {
        "tags": [
          "Data Access"
        ],
        "summary": "Get Records (Paginated)",
        "description": "Fetch records with cursor-based pagination, filtering, and sorting. Use for infinite scroll, data tables, or export pipelines.",
        "operationId": "getRecords",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Table name"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Records per page"
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Pagination cursor from previous response"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Column to sort by"
          },
          {
            "name": "sort_dir",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tables/{table}/records/{pk}": {
      "get": {
        "tags": [
          "Data Access"
        ],
        "summary": "Get Record by Primary Key",
        "description": "Fetch a single record by its primary key. Use for detail pages, profile views, or direct lookups.",
        "operationId": "getRecordByPK",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Table name"
          },
          {
            "name": "pk",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Primary key value"
          }
        ],
        "responses": {
          "200": {
            "description": "Record retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Record object with dynamic fields"
                }
              }
            }
          },
          "404": {
            "description": "Record not found"
          }
        }
      }
    },
...
    "/api/search/global": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Global Search (All Tables)",
        "description": "Search across all tables simultaneously with offset pagination. Use for universal search boxes, admin dashboards, or cross-entity discovery.",
        "operationId": "globalSearch",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100
            },
            "description": "Search term"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Results per table"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Skip N results"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Page number (alternative to offset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Global search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tables/{table}/stats": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Table Statistics",
        "description": "Get row counts and sync statistics. Use for dashboard metrics, data quality monitoring, or sync verification.",
        "operationId": "getTableStats",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Table name"
          }
        ],
        "responses": {
          "200": {
            "description": "Table statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStats"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TableSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "data_type": {
                  "type": "string"
                },
                "is_nullable": {
                  "type": "boolean"
                },
                "is_primary_key": {
                  "type": "boolean"
                }
              }
            }
          },
          "primary_key": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "indexes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sortable": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filterable": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "searchable": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clickhouse_search": {
            "type": "boolean"
          }
        }
      },
      "RecordsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "next_cursor": {
            "type": "string"
          },
          "has_more": {
            "type": "boolean"
          },
          "count": {
            "type": "integer"
          },
          "table": {
            "type": "string"
          },
          "cached": {
            "type": "boolean"
          }
        }
      },
...
      "GlobalSearchResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "count": {
                  "type": "integer"
                },
                "total": {
                  "type": "integer"
                },
                "page": {
                  "type": "integer"
                },
                "limit": {
                  "type": "integer"
                },
                "offset": {
                  "type": "integer"
                },
                "total_pages": {
                  "type": "integer"
                },
                "has_next": {
                  "type": "boolean"
                },
                "has_previous": {
                  "type": "boolean"
                },
                "search_engine": {
                  "type": "string"
                }
              }
            }
          },
          "total_results": {
            "type": "integer"
          },
          "tables_searched": {
            "type": "integer"
          },
          "search_time": {
            "type": "integer",
            "description": "Search time in milliseconds"
          },
          "query": {
            "type": "string"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "clickhouse_available": {
            "type": "boolean"
          }
        }
      },
      "TableStats": {
        "type": "object",
        "properties": {
          "table": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "estimated_rows": {
            "type": "integer"
          },
          "clickhouse_indexed": {
            "type": "integer"
          },
          "clickhouse_last_sync": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}